add new func

+++ ft_atoi.c
@@ -3,35 +3,37 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yoelhaim <yoelhaim@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: yoelhaim <yoelhaim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 16:05:45 by yoelhaim          #+#    #+#             */
-/*   Updated: 2021/11/01 16:16:25 by yoelhaim         ###   ########.fr       */
+/*   Updated: 2021/11/02 10:26:17 by yoelhaim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+#include "libft.h"
+
+size_t	ft_atoi(const char *str)
 {
-	int	i;
-	int n;
-	int	r;
-	
+	size_t	i;
+	size_t	n;
+	size_t	r;
+
 	i = 0;
 	n = 1;
 	r = 0;
-	while(str[i] <= 32)
+	while (str[i] <= 32)
 		i++;
-	if(str[i]== '-')
+	if (str[i] == '-')
 	{
 		n = -1;
 		i++;
 	}
 	else if (str[i] == '+')
 		i++;
-	while(str[i] && str[i] >= '0' && str[i] <= '9')
+	while (str[i] && str[i] >= '0' && str[i] <= '9')
 	{
-		r *=10;
-		r += str[i] -48;
+		r *= 10;
+		r += str[i] - 48;
 		i++;
 	}
 	return (r * n);
